2006-09-01  Andy Wingo  <wingo@pobox.com>
* guile/g-wrap/guile/ws/standard.scm (destroy-value-cg): A
non-inlined non-null caller-owned string always comes from an
argument, which is copied from its scheme values. So while from
the perspective of the code that we call, it is a const char*, we
will need to free it at the end. Cast so no warning is emitted.
This is a bit of a hack; the other solution would be to have
everything be a char*. Better to keep the compiler warnings.