g-wrap reference manual updated for texinfo 5.2
Note: it all started because I wanted to get acces to modern procedure
and type/class indexes, as produced by texinfo 5.2.  then I realized
that, appart from having to solve raised bugs compiling with texinfo
5.2, it really could live with a serious lifting :).  I did my best to
improve the presentation and access to the great material the previous
g-wrap manual already had, mainly by restructuring it: it is now
almost identical, in structure, to the guile reference manual.

It is not easy to describe all these modifications in a few words, but
I can try :).

The top level [g-wrap.texi] document now only specify the license
extract dealing with the documentation itself.

The detailed-menu has been completely removed.

A Preface [non numbered] chapter has been added with 2 sections:
'Contributors to this manual' and 'The G-wrap License'.

The previous 'Overview' introduction section text has been moved and
is now the introduction text itself, preceeded by the first sentence
of the README file: it really is much clearer this way, I think.

'Why Create a Wrapper Generator?' becomes the first Introduction
chaper section. It is followed by 'Obtaining and Installing G-wrap' a
section which is new, and ends with the 'A simple example' section.

'Usage' [chapter] has been renamed 'A more detailed example', with its
text being what was precisly named 'A more detailed example' in the
previous version of this manual. The following 2 sections are
unchanged. 'The aggregated Type Qualifier' section has been added [the
original text was there but unused somehow]. This chaper now ends with
a new section which is 'Reporting bugs'.

The 'API Reference' chapter is unchanged.

Appendix 'A GNU Free Documentation License' added.  'Procedure and
Method Index' now precceeds the 'Type and Class Index', as for the
guile reference manual.

* doc/aggregate.texi:
* doc/effective-version.texi:
* doc/fdl.texi:
* doc/install.texi:
* doc/preface.texi:
* doc/report-bugs.texi:
* doc/version.texi: added

* doc/g-wrap.texi: modified as described above.